+++ src/shared/ui/select/data.ts
