+++ src/shared/ui/select/index.ts
@@ -1,0 +1,7 @@
+export * from '@/shared/ui/select/InputFieldSelect'
+
+export * from '@/shared/ui/select/data.ts'
+
+export * from '@/shared/ui/select/select.tsx'
+
+export * from '@/shared/ui/select/type.ts'
