+++ src/shared/ui/textField/index.ts
@@ -1,0 +1,3 @@
+export * from '@/shared/ui/textField/PhoneTextField'
+
+export * from '@/shared/ui/textField/SearchTextField'
